destination_x/requirements.txt
flask
flask_sqlalchemy
flask_migrate
psycopg2-binary
python-dotenv
aiohttp
httpx
beautifulsoup4
readability-lxml
lxml[html_clean]
openai
pycountry
asyncio
trafilatura
requests
pandas
openpyxl
apscheduler
mysql-connector-python
amadeus
Flask[async]